+++ src/main/java/com/stream_pi/server/window/settings/GeneralSettings.java
@@ -29,6 +29,7 @@ import com.stream_pi.util.alert.StreamPiAlertListener;
 import com.stream_pi.util.alert.StreamPiAlertType;
 import com.stream_pi.util.checkforupdates.CheckForUpdates;
 import com.stream_pi.util.checkforupdates.UpdateHyperlinkOnClick;
+import com.stream_pi.util.restart.Restart;
 import com.stream_pi.util.exception.MinorException;
 import com.stream_pi.util.exception.SevereException;
 import com.stream_pi.util.platform.PlatformType;
@@ -75,6 +76,7 @@ public class GeneralSettings extends VBox
     private final Button saveButton;
     private final Button checkForUpdatesButton;
     private final Button factoryResetButton;
+    private final Button restartButton;
     private final LanguageChooserComboBox languageChooserComboBox;
 
     private Logger logger;
@@ -146,6 +148,9 @@ public class GeneralSettings extends VBox
 
         factoryResetButton = new Button(I18N.getString("window.settings.GeneralSettings.factoryReset"));
         factoryResetButton.setOnAction(actionEvent -> onFactoryResetButtonClicked());
+        
+        restartButton = new Button(I18N.getString("window.settings.GeneralSettings.restart"));
+        restartButton.setOnAction(event->restart());
 
         serverNameTextField.setPrefWidth(200);
 
@@ -175,7 +180,8 @@ public class GeneralSettings extends VBox
                 startOnBootHBox,
                 showAlertsPopupHBox,
                 factoryResetButton,
-                checkForUpdatesButton
+                checkForUpdatesButton,
+                restartButton
         );
 
         vbox.prefWidthProperty().bind(widthProperty().subtract(25));
@@ -213,6 +219,11 @@ public class GeneralSettings extends VBox
                 hostServices.showDocument(getURL());
             }
         });
+    }
+    
+    private void restart()
+    {
+        serverListener.restart();
     }
 
     public void loadData() throws SevereException
