--- window/settings/about/ContactTab.java
@@ -1,63 +1,0 @@
-/*
- * Stream-Pi - Free & Open-Source Modular Cross-Platform Programmable Macro Pad
- * Copyright (C) 2019-2021  Debayan Sutradhar (rnayabed),  Samuel Quiñones (SamuelQuinones)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-package com.stream_pi.server.window.settings.about;
-
-import com.stream_pi.server.i18n.I18N;
-import com.stream_pi.util.links.Links;
-import javafx.application.HostServices;
-import javafx.scene.control.Hyperlink;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.layout.VBox;
-
-
-public class ContactTab extends ScrollPane
-{
-    private final HostServices hostServices;
-
-
-    public ContactTab(HostServices hostServices)
-    {
-        this.hostServices = hostServices;
-
-        getStyleClass().add("about_contact_tab_scroll_pane");
-
-        Hyperlink github = new Hyperlink("GitHub");
-        github.setOnAction(event -> openWebpage(Links.getGitHub()));
-
-        Hyperlink discord = new Hyperlink("Discord");
-        discord.setOnAction(event -> openWebpage(Links.getDiscord()));
-
-        Hyperlink website = new Hyperlink(I18N.getString("window.settings.about.ContactTab.website"));
-        website.setOnAction(event -> openWebpage(Links.getWebsite()));
-
-        Hyperlink twitter = new Hyperlink("Twitter");
-        twitter.setOnAction(event -> openWebpage(Links.getTwitter()));
-
-        Hyperlink matrix = new Hyperlink("Matrix");
-        matrix.setOnAction(event -> openWebpage(Links.getMatrix()));
-
-        VBox vBox = new VBox(github, discord, website, twitter, matrix);
-        vBox.setSpacing(10.0);
-
-        setContent(vBox);
-    }
-
-
-    public void openWebpage(String url)
-    {
-        hostServices.showDocument(url);
-    }
-
-}
